Add support for fixed-current battery charging module.
This is the module shown in the Fritzing diagram.

This module has 4 TP4056 modules modified with 2.4K restistors on Rprog/R3 to set current limit around 500mA.
Thus, there is no need for INA219 current monitors.
Instead it uses an Adafruit breakout board for the ADS1015 4-channel 12-bit ADC to measure the voltage on each of 4 batteries on the module.

The loop detects which type of battery module is connected to the currently selected I2C bus, and then either operates the 4 INA219 modules, or the single ADS1015 4-channel module. It does this by attempting to transmit to the known I2C address of the ADS1015 module (0x48) and reading the error code.

Unfortunately this method of probing for a given I2C address does not give a reliable error code with the standard Arduino Wire library. So Adafruit's TinyWireM library, specifically written for ATTiny[85] modules like Trinket, is used.

And that breaks the INA219 library which is expecting the regular Wire library. So this currenlty uses the big hammer approach of including modified Adafruit_INA219 library files in this project. Credit to: https://github.com/josefadamcik/INA219PowerMeasurementUnit

The switch to TinyWireM also used up more flash, putting this sketch right at the limit of space available...

I hope to find a better solution...