Take readings once per minute.
This is a basic approach to minimizing processor time.
This way the board will not have to continuously check the second value and cycle for the ~55 seconds that nothing needs to happen.
Assuming there is any optimization around time.sleep(), this should save power and minimize heat build up on the chip affecting the temp readings.