Handle the IC I had to swap in from a knock-off of the Adafruit ADS1015 after frying the original.
The IC I swapped in is from one of the skinny purple breakout boards available on Amazon. Branded such as 'Comidox'. The IC itself I suspect is a knock-off of the TI chip. The pinouts appear to be identical but if you use the Adafruit library to ask for single-ended voltage at index 0, you get the value for whatever is connected to A3. So this change remaps things and explains what's going on... At least until I can get some legit ICs or more of the Adafruit breakout board.

Also included in this change is test sketch that works on the Mega.