Update to previous fix for mixed up index to analog input values.
The chip isn't a knock-off as I suspected, in fact it appears to be the higher precision 16-bit ADS1115 instead of the 12-bit ADS1015. So the cheap ADC breakout boards were misadvertised, and the fix is to simply use the other constructor in the Adafruit_ADS1X15 library.

Added `#define ADC1115` to toggle between the two chips. Comment out to use ADS1015, uncommment to use ADC1115.