Reduce flash space needed to detect empty battery slots.
The main change is to the order of checking module type and reading values.

Whereas the previous order checked module type and then did a module-specific iteration across each slot, the new order checks for module type, stores it, then uses a common iteration loop that refers back to the module type as needed.

The result is elimination of LCD printing code and significant flash savings (5020 out of 5310 now).